Only output to istio-owned chains
